requirements.txt
selenium==3.141.0
pandas==1.2.3
gspread==4.0.1
numpy==1.19.2
XlsxWriter==1.2.9
xlrd==1.0.0
openpyxl==3.0.7
django-environ==0.5.0
